Fix demo app unit tests
Now, running
g test angular-cesium-demo will result in a passing test suite.

Added scripts section to the test config in the angular.json file. This avoids the issue where the Cesium
variable is undefined in the tests.

Removed obsolete tests from the app.component.spec.ts file because the elements that they were testing
no longer exist in the AppComponent.

Updated the README file to indicate to future users of the library that they need to include the Cesium
scripts in the test config in order for the tests to know about the Cesium global var.